Analyzed and tried to fix this issue. But I cannot yet find the true issue, in order to no longer delay the next version of YAT I added this bug to the limitations:
USB Ser/HID allows to detect and select multiple devices of the same VID and PID as long as they have a different serial number, however, it is not yet possible to properly communicate with them simultaneously, this could be a bug or a limitation of the Win32 HID.dll (bug #3123900>#169)

To be further analyzed by trying it with other libraries/tools. Can any tool deal with the issue? Or is a limitation of HID.dll for file based serial profiles in the end? Can hardly believe so, but...